Implement automation mechanism
An external script can be used to automate the scheduling.

An event manager receives events from processes, pages and IO queue,
and dispatches them to the external script.

The external script then processes the events and returns a list of
actionable events (simulating clicks).

The game scene dispatches the received events to the respective receivers.

Main changes are:
- event manager;
- changes to provide events to event manager;
- compile and execute an external script;
- _on_click/_onClick is now public;
- game's scene dispatches events from script before calling the update function

An example automated script is provided, mostly to explain the
exposed API, it has a (not so) simple example implementation.